+++ packages/v4/gatsby-config.js
@@ -36,7 +36,7 @@ module.exports = {
         context: 'org',
         hiddenPages: ['withOuia'], // By title
         showBanner: true, // PF3 banner
-        showGdprBanner: true, // GDPR banner
+        showGdprBanner: false, // GDPR banner
         sideNav: {
           core: [
             { section: 'overview' },