+++ packages/v4/src/pages/get-in-touch.js
@@ -5,11 +5,8 @@ import { Grid, GridItem, PageSection, PageSectionVariants, Split, SplitItem, Tit
 import "./get-in-touch.css";
 
 const GetInTouch = ({ location }) => {
-  if (typeof window !== 'undefined') {
-    require('../qualtrics.js');
-  }
   return (
-  <SideNavLayout location={location} hideSideNav showGdprBanner={true} pageTitle="Get in touch">
+  <SideNavLayout location={location} hideSideNav pageTitle="Get in touch">
     <PageSection
       variant={PageSectionVariants.light}
       className="ws-section ws-get-in-touch"