+++ packages/v4/CHANGELOG.md
@@ -3,6 +3,121 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.52.26](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.25...patternfly-org-4@3.52.26) (2020-04-17)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.25](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.24...patternfly-org-4@3.52.25) (2020-04-16)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.24](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.23...patternfly-org-4@3.52.24) (2020-04-16)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.23](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.22...patternfly-org-4@3.52.23) (2020-04-13)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.22](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.21...patternfly-org-4@3.52.22) (2020-04-09)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.21](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.20...patternfly-org-4@3.52.21) (2020-04-09)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.20](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.19...patternfly-org-4@3.52.20) (2020-04-03)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.19](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.18...patternfly-org-4@3.52.19) (2020-04-02)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.18](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.17...patternfly-org-4@3.52.18) (2020-04-02)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.17](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.16...patternfly-org-4@3.52.17) (2020-04-01)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.16](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.15...patternfly-org-4@3.52.16) (2020-04-01)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.15](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.14...patternfly-org-4@3.52.15) (2020-04-01)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.14](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.13...patternfly-org-4@3.52.14) (2020-03-31)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [3.52.13](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.12...patternfly-org-4@3.52.13) (2020-03-31)
+
+
+### Bug Fixes
+
+* **docs:** fix broken datatoolbar links in bulk selection docs ([#1751](https://github.com/patternfly/patternfly-org/issues/1751)) ([bc732a4](https://github.com/patternfly/patternfly-org/commit/bc732a48b37a342f91628e0b4a1f6e8d0dabef38))
+
+
+
+
+
 ## [3.52.12](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@3.52.11...patternfly-org-4@3.52.12) (2020-03-20)
 
 **Note:** Version bump only for package patternfly-org-4